Game is working
All win conditions have checks now, just trying to spruce up the way the game looks, clean up the code for readability, and then make an example AI to play against.

FOR THE AI TOURNAMENT SYSTEM:

Make it so that each script is added to a List in the first tournament scene and then randomly choose a tournament bracket and make the AI fight against each other. then do each layer in the bracket and get up to the top.

USE A BINARY TREE FOR THAT.