Update LDAdvancedAI.cs
Just changed the drawDirections method to only draw a specific direction and made the getDirections method only return a list of slots in one specific direction, not trying to return all 4. This way we can just check the specific directions we want in a specific order and then place coins accordingly.