Updated GridGame.cs, AIPlayer.cs, LDAIExample.cs, and LDAdvancedAI.cs
Added an OnGameStart event so the AI can tell when the game starts and decide if they need to move or not, also fixed a ton of small bugs regarding the AI and the GridGame interacting. Most of them were the AI playing when it wasn't their turn or the Player playing when it wasn't their turn. Going to continue working on LDAdvancedAI.cs so I can test the limits of the program.