Updated GameGrid.cs
Changed it so that now there are UI's that display info on the AI and there is UI input for slowMo toggle and the amount of time it waits between turns before calling switchTurn. I also converted previous references I had to set in every scene to just get them at runtime. I know it isn't exactly the best way to do it typically but I am doing it so that when people clone the repo references won't break in the inspector, it'll also help if we procedurally setup scenes for a tournament like scenario.